so much of this needs to be organized
